requirements.txt
flask
gunicorn
transformers
Pillow
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
